cut/paste AccessorProperty and PandasDelegate to core.accessor
